refactor(1682 #3): split root.py into domain section modules + wire imports
Commit 2 of the config split (#1682 #3): decompose the all-in-root checkpoint into
domain-concern section modules (grouped per the plan, NOT yaml-key shape):

  root.py       ReynConfig + model/tier fields + model_class_for (+ R1 concrete
                section-config imports so get_type_hints resolves)
  loader.py     load_config / _merge / _load_yaml / shape-wiring (yaml-coupled core)
  chat.py       Reasoning/Chat/Loop/Compaction/Timeout/OnLimit/Safety (+ cost builders)
  embedding.py  Embedding/SkillSearch/ActionRetrieval
  media.py      Voice/Multimodal/Web/WebFetch
  execution.py  Plan/SkillResume/SelfImprovement/TimeTravel/ToolUse
  infra.py      Agent/Auth/Sandbox/Events/Eval/Cron/Python

The 2.8k-line god-module → 7 modules (largest 725L, no god-module). `__init__`
re-exports the union (mechanical namespace copy → R4-proof).

Hazards handled:
- R1 (silent forward-ref): root.py CONCRETE-imports every section config it
  references; GATE verified GREEN — test_config_schema_introspection +
  test_config_mirror_coverage_1056 (get_type_hints(ReynConfig) resolves, no
  dropped schema fields).
- R2 (cycles): import DAG cycle-checked acyclic — root→sections, execution→chat,
  loader→{root,sections}. Broke 3 would-be cycles by moving the helper to its
  CONSUMER section (else section↔loader / section↔root): cost builders
  (_build_cost_limit/_build_cost_config) → chat; _default_agent_id → infra (placed
  before AgentConfig, its default_factory consumer); _empty_external_transports →
  root. reyn.* imports stay lazy/function-local (only module-level external dep is
  reyn.budget.budget — verified no reverse import).
- Future-redesign seam: _merge + load_config shape-wiring stay centralized in
  loader.py (a future reyn.yaml-format redesign touches loader.py only).

test_routerloop_convergence_entrypoint_parity_1092: the op-loop-gate
leaf-reachability invariant flagged config/root.py because the split separated the
ReynConfig field DECLARATION (root) from load_config's construction (loader). The
gate still propagates (loader). Refined _references_gate to not treat the dataclass
field-DECLARATION (AnnAssign target) as a drop-hazard — it is the gate's home.

Refs #1682
